% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R, R/oxy-extract_functions.R
\name{extract_functions}
\alias{extract_functions}
\title{FUNCTION_TITLE}
\usage{
extract_functions(path)

extract_functions(path)
}
\arguments{
\item{path}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION

OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

FUNCTION_DESCRIPTION
}
\details{
DETAILS

DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{map}}
\code{\link[vctrs]{vec_slice}}

\code{\link[purrr]{map}}
\code{\link[vctrs]{vec_slice}}
}
