% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_find_fun.R
\name{file_find_fun}
\alias{file_find_fun}
\title{Find which files call a given function}
\usage{
file_find_fun(file, fun)
}
\arguments{
\item{file}{file path}

\item{fun}{function name like "source" or "library"}
}
\value{
dataframe where functions are called
}
\description{
Find which files call a given function
}
