% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_sources.R
\name{file_library}
\alias{file_library}
\title{Find where \code{library()} is called}
\usage{
file_library(file)
}
\arguments{
\item{file}{file path}
}
\value{
dataframe of places where "library()" is called
}
\description{
Find where \code{library()} is called
}
