% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-sources.R
\name{file_sources}
\alias{file_sources}
\title{Which files call "source"?}
\usage{
file_sources(file)
}
\arguments{
\item{file}{file path}
}
\value{
dataframe of which files use source
}
\description{
Which files call "source"?
}
\examples{
\dontrun{
library(purrr)
r_source_files <- list.files(path = "R/",
                             full.names = TRUE,
                             pattern = ".R$")

r_source_files

which_code_needs_what <- map_dfr(r_source_files, file_sources)
which_code_needs_what
}
}
