% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_name.R
\name{fun_name}
\alias{fun_name}
\title{Extract function name from xml}
\usage{
fun_name(fun_text)
}
\arguments{
\item{fun_text}{list of extracted functions}
}
\value{
names of functions
}
\description{
For use inside other functions
}
\note{
internal
}
