% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_names.R
\name{function_names}
\alias{function_names}
\title{Extract function names from a provided path}
\usage{
function_names(path)
}
\arguments{
\item{path}{file path}
}
\value{
character vector of function names
}
\description{
Extract function names from a provided path
}
