% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n-lines.R
\name{n_lines}
\alias{n_lines}
\alias{n_lines_df}
\title{The number of lines in code}
\usage{
n_lines(path)

n_lines_df(path)
}
\arguments{
\item{path}{path to file}
}
\value{
number of lines of code
}
\description{
The number of lines in code
}
\examples{
\dontrun{
n_lines("R/functions.R")
files <- list.files(path = "R/",
                    full.names = TRUE)
n_lines_df(dir_ls(path = "R/"))
}
}
