% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_as_xml.R
\name{parse_as_xml}
\alias{parse_as_xml}
\title{Parse an R file as parse data in XML}
\usage{
parse_as_xml(path)
}
\arguments{
\item{file}{path}
}
\value{
xml
}
\description{
Parse an R file as parse data in XML
}
