% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_xml.R
\name{pretty_print_xml}
\alias{pretty_print_xml}
\title{Nicely print out xml of R script}
\usage{
pretty_print_xml(path)
}
\arguments{
\item{path}{file path}
}
\value{
XML printed nicely
}
\description{
Handy for inspecting full XML path of an R script
}
\examples{
\dontrun{
if(interactive()){
pretty_print_xml("inst/test-fun.R")
 }
}
}
