% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_libraries.R
\name{script_libs}
\alias{script_libs}
\alias{script_libs_df}
\title{Extract the libraries used in a set of scripts}
\usage{
script_libs(path)

script_libs_df(path)
}
\arguments{
\item{path}{path to file}
}
\value{
dataframe of libraries used
}
\description{
Extract the libraries used in a set of scripts
}
\examples{
\dontrun{
script_libs("R/functions.R")
script_libs("R/counterfactuals.R")
script_libs_df(dir_ls(path = "R/"))
}
}
