% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_expand.R
\name{seq_expand}
\alias{seq_expand}
\title{Expand a vector to go from first to the second vector}
\usage{
seq_expand(vec)
}
\arguments{
\item{vec}{vector of length two}
}
\value{
sequence
}
\description{
Expand a vector to go from first to the second vector
}
\note{
internal
}
