% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_mat.R
\name{split_mat}
\alias{split_mat}
\title{Split a matrix into lists}
\usage{
split_mat(x)
}
\arguments{
\item{x}{matrix}
}
\value{
list of different matrices
}
\description{
Split a matrix into lists
}
\note{
interal
}
