% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_end.R
\name{sym_end}
\alias{sym_end}
\title{Get where symbol starts in the data}
\usage{
sym_end(xml)
}
\arguments{
\item{xml}{xml parse data}
}
\value{
symbol end position
}
\description{
Get where symbol starts in the data
}
\note{
internal
}
