% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym_start.R
\name{sym_start}
\alias{sym_start}
\title{Get where symbol ends in the data}
\usage{
sym_start(xml)
}
\arguments{
\item{xml}{xml parse data}
}
\value{
symbol start position
}
\description{
Get where symbol ends in the data
}
\note{
internal
}
