% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-conflicted.R
\name{use_conflicted}
\alias{use_conflicted}
\title{Use conflicted and set up a basic use of it}
\usage{
use_conflicted(file = "conflicts.R")
}
\arguments{
\item{file}{file to write to, default is "conflicts.R"}
}
\value{
nothing, it writes a file
}
\description{
Use conflicted and set up a basic use of it
}
\examples{
\dontrun{
use_conflicted
}
}
