% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_packages.R
\name{write_packages}
\alias{write_packages}
\title{Writes a packages.R file based on dependencies}
\usage{
write_packages(dir = "R/", file = "packages.R")
}
\arguments{
\item{dir}{directory to search through. Default is "R/"}

\item{file}{file to write to - default is "packages.R"}
}
\value{
nothing, it writes a file
}
\description{
Writes a packages.R file based on dependencies
}
\examples{
\dontrun{
write_packages()
}
}
